+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- The engine to the .deb package.
+
 ### Fixed
 
 - Don't allow the Heat Map when the game is over.