+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- The engine to the .deb package.
+- The engine and the server to the .deb package.
 
 ### Changed