+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.3.0] - 2025-11-24
+
 ### Added
 
 - hnefatafl-client: the option to only use ASCII characters to do rendering.
@@ -507,7 +509,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - hnefatafl-client: Make users and games scrollable.
 - Make all the features dependencies.
 
-[unreleased]: https://github.com/dcampbell24/hnefatafl/compare/v4.2.2...main
+[unreleased]: https://github.com/dcampbell24/hnefatafl/compare/v4.3.0...main
+[4.3.0]: https://github.com/dcampbell24/hnefatafl/compare/v4.2.2...v4.3.0
 [4.2.2]: https://github.com/dcampbell24/hnefatafl/compare/v4.2.0...v4.2.2
 [4.2.0]: https://github.com/dcampbell24/hnefatafl/compare/v4.1.1...v4.2.0
 [4.1.1]: https://github.com/dcampbell24/hnefatafl/compare/v4.1.0...v4.1.1